+++ cdp/network/types.go
@@ -283,7 +283,7 @@ func (t *ResourcePriority) UnmarshalJSON(buf []byte) error {
 	return easyjson.Unmarshal(buf, t)
 }
 
-// Request hTTP request data.
+// Request HTTP request data.
 type Request struct {
 	URL              string                    `json:"url"`                        // Request URL.
 	Method           string                    `json:"method"`                     // HTTP request method.
@@ -378,10 +378,10 @@ func (t *BlockedReason) UnmarshalJSON(buf []byte) error {
 	return easyjson.Unmarshal(buf, t)
 }
 
-// Response hTTP response data.
+// Response HTTP response data.
 type Response struct {
 	URL                string           `json:"url"`                          // Response URL. This URL can be different from CachedResource.url in case of redirect.
-	Status             float64          `json:"status"`                       // HTTP response status code.
+	Status             int64            `json:"status"`                       // HTTP response status code.
 	StatusText         string           `json:"statusText"`                   // HTTP response status text.
 	Headers            Headers          `json:"headers"`                      // HTTP response headers.
 	HeadersText        string           `json:"headersText,omitempty"`        // HTTP response headers text.
@@ -408,7 +408,7 @@ type WebSocketRequest struct {
 
 // WebSocketResponse webSocket response data.
 type WebSocketResponse struct {
-	Status             float64 `json:"status"`                       // HTTP response status code.
+	Status             int64   `json:"status"`                       // HTTP response status code.
 	StatusText         string  `json:"statusText"`                   // HTTP response status text.
 	Headers            Headers `json:"headers"`                      // HTTP response headers.
 	HeadersText        string  `json:"headersText,omitempty"`        // HTTP response headers text.
